### 2025-04-02 #7981 @arfon
@TimoDiepers  – - thanks for your submission to JOSS. We're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.
