### 2025-04-10 208628b
Update dependency bw2io to v0.9.9
