### 2025-04-29 1c460a1
Update dependency bw2data to v4.4.4
