### 2025-06-26 859c82f
Update dependency bw2io to v0.9.11
