### 2025-08-06 bbf6124
add CI/CD tests badge to readme
