### 2025-08-06 3459950
Rename CI/CD tests workflow
