### 2025-08-10 6e7e828
Update dependency scipy to <1.17
