### 2025-08-20 e31cbd6
Complete Panel web application with modular structure and URL routing
Co-authored-by: TimoDiepers <90762029+TimoDiepers@users.noreply.github.com>
