+++ app/results/__init__.py
@@ -1,0 +1,1 @@
+"""Results module for bw_timex web application."""
