+++ app/app.py
@@ -1,0 +1,347 @@
+"""
+Main Panel application for bw_timex with URL-based routing.
+
+This application provides a web interface for time-explicit Life Cycle Assessment
+using the bw_timex library with proper URL routing and modular structure.
+"""
+
+import panel as pn
+from typing import Dict, Callable
+import os
+from pathlib import Path
+
+# Import view modules
+try:
+    from .modeling import calculation_setup, timeline_view
+    from .results import lci_results, lcia_results
+except ImportError:
+    # Handle relative imports when running directly
+    import sys
+    sys.path.append(str(Path(__file__).parent))
+    from modeling import calculation_setup, timeline_view
+    from results import lci_results, lcia_results
+
+# Enable Panel extensions - simplified
+pn.extension('tabulator')
+
+# Load CSS
+css_path = Path(__file__).parent / "static" / "styles.css"
+if css_path.exists():
+    with open(css_path, 'r') as f:
+        css_content = f.read()
+    pn.config.raw_css = [css_content]
+
+# Configure Panel
+pn.config.sizing_mode = 'stretch_width'
+
+
+def create_divider():
+    """Create a simple divider."""
+    return pn.pane.HTML('<hr style="margin: 20px 0; border: 1px solid #ddd;">')
+
+
+class TimexApp:
+    """Main application class with URL routing."""
+    
+    def __init__(self):
+        self.routes = self._setup_routes()
+        self.current_view = None
+        self.current_path = '/'  # Track current path
+        
+    def _setup_routes(self) -> Dict[str, Callable]:
+        """Setup URL routes mapping to view functions."""
+        return {
+            '/': self.home_view,
+            '/modeling/calculation-setup': calculation_setup.view,
+            '/modeling/timeline': timeline_view.view,
+            '/results/lci': lci_results.view,
+            '/results/lcia': lcia_results.view,
+        }
+    
+    def home_view(self):
+        """Home page view."""
+        return pn.Column(
+            pn.pane.Markdown("# Welcome to bw_timex Web Interface"),
+            pn.pane.Markdown("""
+            This is the web interface for **bw_timex** - Time-explicit Life Cycle Assessment.
+            
+            ## Features
+            - **Modeling**: Set up calculations and configure timelines
+            - **Results**: View LCI and LCIA results
+            
+            ## Navigation
+            Use the menu above to navigate between different sections:
+            - **Modeling**: Define your time-explicit LCA setup
+            - **Results**: Analyze your calculation results
+            
+            ## Getting Started
+            1. Go to [Calculation Setup](/modeling/calculation-setup) to configure your analysis
+            2. Set up your [Timeline](/modeling/timeline) 
+            3. View your [Results](/results/lci) once calculations are complete
+            """),
+            sizing_mode='stretch_width'
+        )
+    
+    def create_navbar(self):
+        """Create navigation bar with menu items."""
+        # Simple navigation buttons that work with location routing
+        nav_buttons = []
+        
+        # Home button
+        home_btn = pn.widgets.Button(
+            name="🏠 Home",
+            button_type='primary',
+            width=120,
+            margin=(5, 5)
+        )
+        home_btn.on_click(lambda event: self.navigate_to('/'))
+        nav_buttons.append(home_btn)
+        
+        # Modeling section
+        calc_setup_btn = pn.widgets.Button(
+            name="📊 Calculation Setup", 
+            button_type='light',
+            width=160,
+            margin=(5, 5)
+        )
+        calc_setup_btn.on_click(lambda event: self.navigate_to('/modeling/calculation-setup'))
+        nav_buttons.append(calc_setup_btn)
+        
+        timeline_btn = pn.widgets.Button(
+            name="⏱️ Timeline",
+            button_type='light', 
[…]
