### 2025-09-19 3385506
Update dependency bw2calc to v2.1.1
