+++ tests/fixtures/dynamic_biomatrix_db_fixture.py
@@ -8,7 +8,6 @@ from bw_temporalis import TemporalDistribution
 @pytest.fixture
 @bw2test
 def dynamic_biosphere_matrix_db():
-    bd.projects.set_current("__test_dynamic_biosphere_matrix__")
     bd.Database("bio").write(
         {
             ("bio", "CO2"): {
