### 2025-09-29 161bd69
logging for utils
