+++ bw_timex/dynamic_biosphere_builder.py
@@ -135,7 +135,7 @@ class DynamicBiosphereBuilder:
         dynamic_biosphere_matrix : scipy.sparse.csr_matrix
             A sparse matrix with the dimensions (bio_flows at a specific time step) x (processes).
         temporal_market_lcis : dict
-            A dictionary containing the disaggregated LCI's of the temporal markets, 
+            A dictionary containing the disaggregated LCI's of the temporal markets,
             with the time-mapped-activity id as key.
         """
 
