### 2025-09-29 2184b03
first teaching notebook draft
