### 2025-09-29 8a9c3d7
move waterfall plot legend outside
