### 2025-09-29 c266587
change order of TimexLCA init functions so that the node_collection is correct
