### 2025-09-29 e6df376
comment out mac specific packages
