+++ environment.yml
@@ -6,6 +6,6 @@ channels:
   - diepers
 dependencies:
   - bw_timex>=0.3.1
-  - brightway25_nosolver # uncomment for macOS ARM machines
-  - scikit-umfpack # uncomment for macOS ARM machines
-  - numpy==1.24 # uncomment for macOS ARM machines
+  # - brightway25_nosolver # uncomment for macOS ARM machines
+  # - scikit-umfpack # uncomment for macOS ARM machines
+  # - numpy==1.24 # uncomment for macOS ARM machines
