### 2025-09-30 06b304e
update requirements pin minimum versions for bw
