+++ environment.yml
@@ -8,4 +8,3 @@ dependencies:
  - bw_timex>=0.3
  - brightway25_nosolver # uncomment for macOS ARM machines
  - scikit-umfpack # uncomment for macOS ARM machines
- - numpy==1.24 # uncomment for macOS ARM machines
