### 2025-09-30 0a14ba7
remove numpy pin for mac osx arm installation imstruction
