### 2025-10-06 a662cd5
add interactive td drafting widget
