+++ bw_timex/utils.py
@@ -2,12 +2,27 @@ from datetime import datetime, timedelta
 from typing import Callable, List, Optional, Union
 
 import matplotlib.pyplot as plt
+import numpy as np
 import pandas as pd
 from bw2data.backends import ActivityDataset as AD
 from bw2data.backends.proxies import Exchange
 from bw2data.backends.schema import ExchangeDataset
 from bw2data.errors import MultipleResults, UnknownObject
-from bw_temporalis import TemporalDistribution
+from bw_temporalis import TemporalDistribution, easy_timedelta_distribution
+from IPython.display import Javascript, display
+from ipywidgets import (
+    Button,
+    Dropdown,
+    FloatSlider,
+    HBox,
+    IntSlider,
+    IntText,
+    Label,
+    Output,
+    Textarea,
+    ToggleButtons,
+    VBox,
+)
 from loguru import logger
 
 time_res_mapping_strftime = {
@@ -38,8 +53,9 @@ def extract_date_as_integer(dt_obj: datetime, time_res: Optional[str] = "year") 
 
     """
     if time_res not in time_res_mapping_strftime:
+        available = list(time_res_mapping_strftime.keys())
         raise ValueError(
-            f"Invalid time_res: '{time_res}'. Please choose from: {list(time_res_mapping_strftime.keys())}."
+            f"Invalid time_res: '{time_res}'. Please choose from: {available}."
         )
     formatted_date = dt_obj.strftime(time_res_mapping_strftime[time_res])
     date_as_integer = int(formatted_date)
@@ -49,8 +65,9 @@ def extract_date_as_integer(dt_obj: datetime, time_res: Optional[str] = "year") 
 
 def extract_date_as_string(timestamp: datetime, temporal_grouping: str) -> str:
     """
-    Extracts the grouping date as a string from a datetime object, based on the chosen temporal grouping.
-    e.g. for `temporal_grouping` = 'month', and `timestamp` = 2023-03-29T01:00:00, it extracts the string '202303'.
+    Extracts the grouping date as a string from a datetime object, based on the chosen temporal
+    grouping. E.g. for `temporal_grouping` = 'month', and `timestamp` = 2023-03-29T01:00:00, it
+    extracts the string '202303'.
 
 
     Parameters
@@ -453,3 +470,249 @@ def add_temporal_distribution_to_exchange(
     exchange["temporal_distribution"] = temporal_distribution
     exchange.save()
     logger.info(f"Added temporal distribution to exchange {exchange}.")
+
+
+def interactive_td_widget():
+    """
+    Create an interactive ipywidget for drafting temporal distributions and copying them to the
+    clipboard.
+
+    For use in jupyter notebooks.
+
+    Returns
+    -------
+    ipywidgets.VBox
+        Interactive widget for drafting temporal distributions.
+    """
+    # ---------- Controls ----------
+    mode = ToggleButtons(
+        options=["Generator", "Manual"], value="Generator", description="Mode"
+    )
+
+    # Generator controls
+    start = IntText(value=0, description="start")
+    end = IntText(value=10, description="end")
+    resolution = Dropdown(
+        options=[("Years", "Y"), ("Months", "M"), ("Days", "D")],
+        value="Y",
+        description="resolution",
+    )
+    steps = IntSlider(
+        value=10, min=2, max=20, step=1, description="steps", continuous_update=False
+    )
+    kind = ToggleButtons(
+        options=["uniform", "triangular", "normal"], value="uniform", description="kind"
+    )
+    # Give wide initial bounds; we'll override on kind changes
+    param = FloatSlider(
+        value=1.0,
+        min=0.01,
+        max=50.0,
+        step=0.01,
+        description="param",
[…]
