+++ pyproject.toml
@@ -102,7 +102,7 @@ formats = "bdist_wheel"
 
 [tool.flake8]
 # Some sane defaults for the code style checker flake8
-max_line_length = 88
+max_line_length = 100
 extend_ignore = ["E203", "W503"]
 # ^  Black-compatible
 #    E203 and W503 have edge cases handled by black
