### 2025-10-06 22a58eb
add ipython and ipywidgets deps
