+++ pyproject.toml
@@ -43,6 +43,8 @@ dependencies = [
     "matplotlib",
     "seaborn",
     "loguru",
+    "ipython",
+    "ipywidgets",
 ]
 
 [project.urls]
