+++ bw_timex/utils.py
@@ -1,3 +1,4 @@
+import json
 from datetime import datetime, timedelta
 from typing import Callable, List, Optional, Union
 
@@ -18,6 +19,7 @@ from ipywidgets import (
     IntSlider,
     IntText,
     Label,
+    Layout,
     Output,
     Textarea,
     ToggleButtons,
@@ -527,19 +529,57 @@ def interactive_td_widget():
         value="Y",
         description="resolution",
     )
-    dates_text = Textarea(value="0, 2, 4, 6, 8, 10", description="dates")
-    amounts_text = Textarea(value="0.1, 0.1, 0.2, 0.2, 0.2, 0.2", description="amounts")
+    dates_text = Textarea(
+        value="0, 2, 4, 6, 8, 10",
+        description="dates",
+        layout=Layout(width="100%", min_height="70px"),
+    )
+    amounts_text = Textarea(
+        value="0.1, 0.1, 0.2, 0.2, 0.2, 0.2",
+        description="amounts",
+        layout=Layout(width="100%", min_height="70px"),
+    )
+
+    for widget in (start, end, resolution, steps, param, manual_unit):
+        widget.style.description_width = "initial"
+
+    steps.layout = Layout(width="220px")
+    param.layout = Layout(width="220px")
+    start.layout = Layout(width="160px")
+    end.layout = Layout(width="160px")
+    resolution.layout = Layout(width="180px")
+    manual_unit.layout = Layout(width="220px")
 
     copy_btn = Button(description="Copy TD code", button_style="success")
+    copy_import_btn = Button(description="Copy TD + imports", button_style="")
+    copy_btn.layout = Layout(width="160px")
+    copy_import_btn.layout = Layout(width="200px")
     status = Label(value="")
-    plot_out = Output()
+    status.layout = Layout(margin="0 0 0 8px")
+    plot_out = Output(layout=Layout(width="100%"))
 
     # ---------- Helpers ----------
-    def _parse_num_list(txt):
+    def _parse_num_list(txt: str, label: str) -> List[float]:
         parts = [p for p in txt.replace(",", " ").split() if p]
-        return [
-            float(p) if (("." in p) or ("e" in p.lower())) else int(p) for p in parts
-        ]
+        if not parts:
+            raise ValueError(f"{label} cannot be empty.")
+
+        values = []
+        for p in parts:
+            try:
+                if ("." in p) or ("e" in p.lower()):
+                    values.append(float(p))
+                else:
+                    values.append(int(p))
+            except ValueError as exc:
+                raise ValueError(f"Could not parse '{p}' in {label}.") from exc
+        return values
+
+    def _format_number(value: float) -> str:
+        as_float = float(value)
+        if np.isfinite(as_float) and as_float.is_integer():
+            return str(int(as_float))
+        return (f"{as_float:.6f}").rstrip("0").rstrip(".")
 
     def _make_td_generator():
         return easy_timedelta_distribution(
@@ -552,12 +592,16 @@ def interactive_td_widget():
         )
 
     def _make_td_manual():
-        d = _parse_num_list(dates_text.value)
-        a = _parse_num_list(amounts_text.value)
+        d = _parse_num_list(dates_text.value, "dates")
+        a = _parse_num_list(amounts_text.value, "amounts")
         if len(d) != len(a):
             raise ValueError("dates and amounts must have the same length.")
+        if not d:
+            raise ValueError("Provide at least one date and amount.")
         date = np.array(d, dtype=f"timedelta64[{manual_unit.value}]")
         amount = np.array(a, dtype=float)
+        if np.any(np.isnan(amount)):
+            raise ValueError("Amounts must be numeric values.")
[…]
