### 2025-10-06 0e652b2
more sensible default values
