### 2025-10-07 d7a6a7b
pin astroid version for docs
