+++ .github/workflows/python-package-deploy.yml
@@ -19,7 +19,7 @@ jobs:
     - name: Set up Python 3.11
       uses: actions/setup-python@v6
       with:
-        python-version: "3.11"
+        python-version: "3.14"
     - name: Install pypa/build
       run: >-
         python -m
