### 2025-10-08 8b9fabf
make negative start/end value work
