### 2025-10-08 35518c0
improve logging
