### 2025-10-08 b87d55a
Fix matplotlib plot duplication in interactive_td_widget by closing figures
Co-authored-by: TimoDiepers <90762029+TimoDiepers@users.noreply.github.com>
