### 2025-10-08 4ae64a3
Add interactive_dynamic_lcia_widget function to utils
Co-authored-by: TimoDiepers <90762029+TimoDiepers@users.noreply.github.com>
