+++ bw_timex/utils.py
@@ -5,6 +5,7 @@ from typing import Callable, List, Optional, Union
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
+import seaborn as sb
 from bw2data.backends import ActivityDataset as AD
 from bw2data.backends.proxies import Exchange
 from bw2data.backends.schema import ExchangeDataset
@@ -828,5 +829,225 @@ def interactive_td_widget():
         refresh_preview()
 
     mode.observe(_mode_refresh, names="value")
+
+    return container
+
+
+def interactive_dynamic_lcia_widget(tlca):
+    """
+    Create an interactive ipywidget for visualizing dynamic LCIA results with interactive controls.
+
+    Takes a TimexLCA object that has already completed lci() and creates an interactive interface 
+    to explore different dynamic LCIA parameters and visualize the results.
+
+    For use in jupyter notebooks.
+
+    Parameters
+    ----------
+    tlca : TimexLCA
+        A TimexLCA object that has already run build_timeline() and lci().
+
+    Returns
+    -------
+    ipywidgets.VBox
+        Interactive widget for dynamic LCIA visualization.
+
+    Raises
+    ------
+    AttributeError
+        If the TimexLCA object hasn't run lci() yet.
+
+    Example
+    -------
+    >>> from bw_timex import TimexLCA
+    >>> from bw_timex.utils import interactive_dynamic_lcia_widget
+    >>> tlca = TimexLCA(demand, method, database_dates)
+    >>> tlca.build_timeline()
+    >>> tlca.lci()
+    >>> interactive_dynamic_lcia_widget(tlca)
+    """
+    # Validation
+    if not hasattr(tlca, "dynamic_inventory"):
+        raise AttributeError(
+            "Dynamic inventory not yet calculated. Call TimexLCA.lci(build_dynamic_biosphere=True) first."
+        )
+
+    # ---------- Controls ----------
+    metric = ToggleButtons(
+        options=["radiative_forcing", "GWP"],
+        value="radiative_forcing",
+        description="Metric",
+    )
+
+    time_horizon = IntSlider(
+        value=100,
+        min=10,
+        max=500,
+        step=10,
+        description="Time horizon (years)",
+        continuous_update=False,
+        style={"description_width": "initial"},
+    )
+
+    fixed_time_horizon = ToggleButtons(
+        options=[("Floating", False), ("Fixed", True)],
+        value=False,
+        description="Time horizon type",
+    )
+
+    # Plot options
+    cumsum = ToggleButtons(
+        options=[("Instantaneous", False), ("Cumulative", True)],
+        value=False,
+        description="Display mode",
+    )
+
+    sum_emissions_within_activity = ToggleButtons(
+        options=[("By flow", False), ("Sum by activity", True)],
+        value=False,
+        description="Emission grouping",
+    )
+
+    sum_activities = ToggleButtons(
+        options=[("By activity", False), ("Sum all", True)],
+        value=False,
+        description="Activity grouping",
+    )
+
+    # Status and output
+    status = Label(value="")
+    plot_out = Output(layout=Layout(width="100%", height="500px"))
+
+    # ---------- Helpers ----------
+    def _run_dynamic_lcia_and_plot(*_):
+        """Run dynamic LCIA with current parameters and plot results."""
+        with plot_out:
+            plot_out.clear_output(wait=True)
+            try:
+                status.value = "⏳ Calculating dynamic LCIA..."
+
+                # Run dynamic LCIA
+                tlca.dynamic_lcia(
+                    metric=metric.value,
+                    time_horizon=time_horizon.value,
+                    fixed_time_horizon=fixed_time_horizon.value,
+                )
+
+                # Create plot
[…]
