### 2025-10-08 e5b1013
Add documentation for interactive_dynamic_lcia_widget
Co-authored-by: TimoDiepers <90762029+TimoDiepers@users.noreply.github.com>
