### 2025-10-08 583bde4
Update CHANGES.md with new feature
Co-authored-by: TimoDiepers <90762029+TimoDiepers@users.noreply.github.com>
