### 2025-10-09 1ede433
Optimize activity name lookups with cached mapping
Co-authored-by: TimoDiepers <90762029+TimoDiepers@users.noreply.github.com>
