+++ CHANGES.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.2] - 2025-10-09
+* Added `bw_timex.utils.interactive_td_widget` for drafting and visualizing temporal distributions in jupyter notebooks
 * Improved caching for activity name lookup
 * Improved logging
 
