+++ CHANGES.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 * Improved caching for activity name lookup
+* Improved logging
 
 ## [0.3.1] - 2025-03-31
 * Fixed an issue with non-unitary production exchanges
