### 2025-10-09 aafe59d
v0.3.2
