+++ bw_timex/__init__.py
@@ -5,4 +5,4 @@ from .matrix_modifier import MatrixModifier
 from .timeline_builder import TimelineBuilder
 from .timex_lca import TimexLCA
 
-__version__ = "0.3.1"
+__version__ = "0.3.2"
