### 2025-10-13 12ffda5
Update dependency bw2data to v4.5.1
