### 2025-10-17 39b926a
Update dependency bw2calc to v2.2
