### 2025-10-17 1f63c39
Add AllEdgeExtractor class for non-priority graph traversal
Co-authored-by: TimoDiepers <90762029+TimoDiepers@users.noreply.github.com>
