+++ bw_timex/edge_extractor.py
@@ -4,6 +4,7 @@ from numbers import Number
 from typing import Callable
 
 import numpy as np
+from bw2calc import LCA
 from bw_temporalis import TemporalDistribution, TemporalisLCA
 
 datetime_type = np.dtype("datetime64[s]")
@@ -225,6 +226,275 @@ class EdgeExtractor(TemporalisLCA):
             return td_consumer
 
         # Else, if both consumer and producer have a td (absolute and relative, respectively) join to TDs
+        if isinstance(td_producer, TemporalDistribution) and isinstance(
+            td_consumer, TemporalDistribution
+        ):
+            if not (td_consumer.date.dtype == datetime_type):
+                raise ValueError(
+                    f"`td_consumer.date` must have dtype `datetime64[s]`, but got `{td_consumer.date.dtype}`"
+                )
+            if not (td_producer.date.dtype == timedelta_type):
+                raise ValueError(
+                    f"`td_producer.date` must have dtype `timedelta64[s]`, but got `{td_producer.date.dtype}`"
+                )
+            date = (
+                td_consumer.date.reshape((-1, 1)) + td_producer.date.reshape((1, -1))
+            ).ravel()
+            amount = np.array(len(td_consumer) * [td_producer.amount]).ravel()
+            return TemporalDistribution(date, amount)
+        else:
+            raise ValueError(
+                f"Can't join TemporalDistribution and something else: Trying with {type(td_consumer.date)} and {type(td_producer.date)}"
+            )
+
+
+class AllEdgeExtractor:
+    """
+    Alternative to EdgeExtractor that traverses the entire supply chain graph
+    without priority-first approach and without calculating LCA impacts.
+    
+    This class provides faster graph traversal for smaller foregrounds by
+    simply convolving temporal distributions of edges and yielding the
+    absolute occurrence of all edges, without the overhead of LCA calculations
+    and prioritization.
+    
+    Unlike EdgeExtractor, this class does NOT inherit from TemporalisLCA.
+    Instead, it builds directly on top of a bw2calc.LCA object.
+    """
+
+    def __init__(
+        self, 
+        lca: LCA,
+        starting_datetime: TemporalDistribution,
+        edge_filter_function: Callable = None,
+        static_activity_indices: set = None,
+        cutoff: float = 1e-9,
+        max_calc: int = 2000,
+        **kwargs
+    ) -> None:
+        """
+        Initialize the AllEdgeExtractor.
+        
+        Parameters
+        ----------
+        lca : LCA
+            A static LCA object that provides the technosphere matrix and
+            activity mappings for the supply chain traversal.
+        starting_datetime : TemporalDistribution
+            The temporal distribution representing the functional unit timing.
+        edge_filter_function : Callable, optional
+            A callable that filters edges based on activity ID. If not provided,
+            a function that always returns False is used.
+        static_activity_indices : set, optional
+            Set of activity indices to skip during traversal.
+        cutoff : float, optional
+            Cutoff threshold for edge amounts. Default is 1e-9.
+        max_calc : int, optional
+            Maximum number of edges to process. Default is 2000.
+        **kwargs : dict
+            Additional keyword arguments for compatibility.
+        """
+        self.lca = lca
+        self.t0 = starting_datetime
+        
+        if edge_filter_function:
+            self.edge_ff = edge_filter_function
+        else:
+            self.edge_ff = lambda x: False
+        
+        self.static_activity_indices = static_activity_indices or set()
+        self.cutoff = cutoff
+        self.max_calc = max_calc
+        
+        # Get mappings from LCA object
+        self.activity_dict = lca.dicts.activity
+        self.technosphere_matrix = lca.technosphere_matrix.tocsr()
+        
[…]
