### 2025-10-17 4ffd898
Add tests and documentation for AllEdgeExtractor, improve demand handling
Co-authored-by: TimoDiepers <90762029+TimoDiepers@users.noreply.github.com>
