### 2025-10-17 c698a53
Enhance documentation with usage examples and drop-in replacement info
Co-authored-by: TimoDiepers <90762029+TimoDiepers@users.noreply.github.com>
