### 2025-10-17 56eaff3
Fix KeyError and type issues in AllEdgeExtractor
- Fix KeyError in build_edge_timeline by using .get() instead of direct dict access
- Convert demand_amount scalar to TemporalDistribution for td_producer to match Edge dataclass type hint
- Add None check after .get() to prevent further errors

Co-authored-by: TimoDiepers <90762029+TimoDiepers@users.noreply.github.com>
