### 2025-10-17 b6a3f13
Expose AllEdgeExtractor option in TimexLCA.build_timeline
- Add use_all_edge_extractor parameter to TimexLCA.build_timeline()
- Pass parameter through to TimelineBuilder
- Update TimelineBuilder to conditionally use AllEdgeExtractor or EdgeExtractor
- Add logging to indicate which extractor is being used
- Update documentation with usage examples for the new parameter
- Add "When to Use" section to help users choose the right extractor

Co-authored-by: TimoDiepers <90762029+TimoDiepers@users.noreply.github.com>
