### 2025-10-17 8421eb2
Fix TypeError in AllEdgeExtractor temporal distribution multiplication
- Add type checking to ensure initial_td is a TemporalDistribution after multiplication
- Add fallback to manually create TemporalDistribution if multiplication returns unexpected type
- Add similar check for distribution after convolution to prevent TypeError
- Fixes "can't multiply sequence by non-int of type 'TemporalDistribution'" error

Co-authored-by: TimoDiepers <90762029+TimoDiepers@users.noreply.github.com>
