### 2025-10-18 2c74876
Add priority_first_traversal parameter to EdgeExtractor
Co-authored-by: TimoDiepers <90762029+TimoDiepers@users.noreply.github.com>
