+++ bw_timex/edge_extractor.py
@@ -6,6 +6,11 @@ from typing import Callable
 import numpy as np
 from bw_temporalis import TemporalDistribution, TemporalisLCA
 
+try:
+    from bw_graph_tools import BreadthFirstGT
+except ImportError:
+    BreadthFirstGT = None
+
 datetime_type = np.dtype("datetime64[s]")
 timedelta_type = np.dtype("timedelta64[s]")
 
@@ -41,7 +46,13 @@ class EdgeExtractor(TemporalisLCA):
     using Brightway Datapackages.
     """
 
-    def __init__(self, *args, edge_filter_function: Callable = None, **kwargs) -> None:
+    def __init__(
+        self,
+        *args,
+        edge_filter_function: Callable = None,
+        priority_first_traversal: bool = True,
+        **kwargs,
+    ) -> None:
         """
         Initialize the EdgeExtractor class and traverses the supply chain using
         functions of the parent class TemporalisLCA.
@@ -52,6 +63,10 @@ class EdgeExtractor(TemporalisLCA):
         edge_filter_function : Callable, optional
             A callable that filters edges. If not provided, a function that always
             returns False is used.
+        priority_first_traversal : bool, optional
+            If True (default), uses priority-first graph traversal with a heap based on
+            cumulative_score. If False, uses breadth-first graph traversal from
+            bw_graph_tools.BreadthFirstGT, which does not require cumulative_score.
         **kwargs : Arbitrary keyword arguments
 
         Returns
@@ -67,14 +82,30 @@ class EdgeExtractor(TemporalisLCA):
         else:
             self.edge_ff = lambda x: False
 
+        self.priority_first_traversal = priority_first_traversal
+
+        # Validate that BreadthFirstGT is available if needed
+        if not self.priority_first_traversal and BreadthFirstGT is None:
+            raise ImportError(
+                "BreadthFirstGT is not available. Please install bw_graph_tools>=0.4 "
+                "or set priority_first_traversal=True."
+            )
+
     def build_edge_timeline(self) -> list:
         """
         Creates a timeline of the edges from the output of the graph traversal.
-        Starting from the edges of the functional unit node, it goes through
-        each node using a heap, selecting the node with the highest impact first.
-        It, then, propagates the TemporalDistributions of the edges from node to
-        node through time using convolution-operators. It stops in case the current edge
-        is known to have no temporal distribution (=leaf) (e.g. part of background database).
+        
+        If priority_first_traversal is True (default):
+            Starting from the edges of the functional unit node, it goes through
+            each node using a heap, selecting the node with the highest impact first.
+        
+        If priority_first_traversal is False:
+            Uses breadth-first graph traversal from bw_graph_tools.BreadthFirstGT,
+            which processes nodes in breadth-first order without using cumulative_score.
+        
+        It propagates the TemporalDistributions of the edges from node to node through time 
+        using convolution-operators. It stops in case the current edge is known to have no 
+        temporal distribution (=leaf) (e.g. part of background database).
 
         Parameters
         ----------
@@ -87,7 +118,23 @@ class EdgeExtractor(TemporalisLCA):
             and consuming node.
 
         """
+        if self.priority_first_traversal:
+            return self._build_edge_timeline_priority_first()
+        else:
+            return self._build_edge_timeline_breadth_first()
 
+    def _build_edge_timeline_priority_first(self) -> list:
+        """
+        Build edge timeline using priority-first graph traversal with heap.
+        
+        This is the original implementation that uses cumulative_score to prioritize
+        nodes with the highest impact first.
+
+        Returns
+        -------
+        list
+            A list of Edge instances.
+        """
         heap = []
         timeline = []
 
[…]
