### 2025-10-18 4073807
Remove visited tracking from breadth-first traversal and add basic tests
Co-authored-by: TimoDiepers <90762029+TimoDiepers@users.noreply.github.com>
