### 2025-10-29 5c19ddd
add citation to readme
