+++ README.md
@@ -35,6 +35,11 @@ Still unsure if you need time-explicit LCA? Overwhelmed by all the options? [Che
 - [Getting Started Tutorial](https://docs.brightway.dev/projects/bw-timex/en/latest/content/getting_started/index.html)
 - [Example Collection](https://docs.brightway.dev/projects/bw-timex/en/latest/content/examples/index.html)
 
+## 📚 Citation
+If `bw_timex` supports your scientific work, please consider citing our companion publications:
+- The conceptual framework and formalization of time-explicit LCA are described in our [methodology paper](https://doi.org/10.1007/s11367-025-02539-3)
+- The implementation of this methodology in `bw_timex` is covered by our [JOSS paper (preprint)](https://github.com/openjournals/joss-papers/blob/joss.07981/joss.07981/10.21105.joss.07981.pdf)
+
 ## 🤝 Contributing
 We welcome contributions! If you have suggestions or want to fix a bug, please:
 - [Open an Issue](https://github.com/brightway-lca/bw_timex/issues)
