+++ joss/paper.bib
@@ -131,9 +131,12 @@
 }
 
 @article{MuellerDiepers:2025,
-  title      = {Time-Explicit Life Cycle Assessment: A Flexible Framework for Coherent Consideration of Temporal Dynamics},
-  author     = {Müller, A and Diepers, T. and Jakobs, A. and Cardellini, G. and von der Assen, N. and Guinée, J. and Steubing, B.},
-  journal = {The International Journal of Life Cycle Assessment},
-  volume = {(submitted)},
-  year       = {2025},
+	title = {Time-explicit life cycle assessment: a flexible framework for coherent consideration of temporal dynamics},
+	issn = {1614-7502},
+	url = {https://doi.org/10.1007/s11367-025-02539-3},
+	doi = {10.1007/s11367-025-02539-3},
+	journal = {The International Journal of Life Cycle Assessment},
+	author = {Müller, Amelie and Diepers, Timo and Jakobs, Arthur and Cardellini, Giuseppe and von der Assen, Niklas and Guinée, Jeroen and Steubing, Bernhard},
+	month = oct,
+	year = {2025},
 }
