+++ joss/paper.bib
@@ -136,7 +136,7 @@
 	url = {https://doi.org/10.1007/s11367-025-02539-3},
 	doi = {10.1007/s11367-025-02539-3},
 	journal = {The International Journal of Life Cycle Assessment},
-	author = {Müller, Amelie and Diepers, Timo and Jakobs, Arthur and Cardellini, Giuseppe and von der Assen, Niklas and Guinée, Jeroen and Steubing, Bernhard},
+	author = {Müller, Amelie and Diepers, Timo and Jakobs, Arthur and Cardellini, Giuseppe and {von der Assen}, Niklas and Guinée, Jeroen and Steubing, Bernhard},
 	month = oct,
 	year = {2025},
 }
