### 2025-11-11 a053d13
Update dependency bw2calc to v2.2.2
