### 2025-11-18 3e2acf7
Update dependency bw2data to v4.5.3
