### 2025-12-01 21c083f
add citation to docs index
