+++ docs/index.md
@@ -18,6 +18,11 @@ You can define temporal distributions for process and emission exchanges, which 
 
 Still unsure if you need time-explicit LCA? Overwhelmed by all the options? [Check out our decision tree](content/decisiontree.md) for some guidance.
 
+## 📚 Citation
+If `bw_timex` supports your scientific work, please consider citing our companion publications:
+- The conceptual framework and formalization of time-explicit LCA are described in our [methodology paper](https://doi.org/10.1007/s11367-025-02539-3)
+- The implementation of this methodology in `bw_timex` is covered by our [JOSS paper (under review)](https://github.com/openjournals/joss-papers/blob/joss.07981/joss.07981/10.21105.joss.07981.pdf)
+
 ## 💬 Support
 If you have any questions or need help, do not hesitate to contact us:
 - Timo Diepers ([timo.diepers@ltt.rwth-aachen.de](mailto:timo.diepers@ltt.rwth-aachen.de))
