### 2025-12-05 bac02d6
Update dependency bw2data to v4.5.4
