### 2025-12-05 97ae8be
Update dependency bw2calc to v2.3
