+++ pyproject.toml
@@ -30,7 +30,7 @@ classifiers = [
 ]
 requires-python = ">=3.11"
 dependencies = [
-    "bw2calc==2.0.dev23",
+    "bw2calc==2.3",
     "bw2data==4.0.dev46",
     "bw2io==0.9.dev31",
     "bw_graph_tools>0.4",
