### 2025-12-13 #9621 @editorialbot
[REVIEW]: bw_timex: A Python Package for Time-Explicit Life Cycle Assessment
**Submitting author:** @TimoDiepers (<a href="http://orcid.org/0009-0002-8566-8618">Timo Diepers</a>)
**Repository:** https://github.com/brightway-lca/bw_timex
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v1.0.0
