### 2025-12-17 6e97ced
Update dependency bw2io to v0.9.12
