### 2025-12-22 brightway-lca/bw_timex#153 @mfastudillo
installation instructions seem to be incomplete
Hi,

I am looking at the [installation instructions](https://docs.brightway.dev/projects/bw-timex/en/latest/content/installation.html) as part of a [review process](https://github.com/openjournals/joss-reviews/issues/9621)

the instructions seem to present two options, either use conda or pip. For conda, it only includes the creation of a virtual environment and the installation of jupyterlab... shouldn't it also include something like :

```bash
conda install diepers::bw_timex
```
to install bw_timex ?
