### 2025-12-22 brightway-lca/bw_timex#154 @mfastudillo
brightway requirements
I am looking at the package requirements, as part of a [review process](https://github.com/openjournals/joss-reviews/issues/9621)

the bw_timex forces the use of particular versions of the main "brightway" packages.

- "bw2calc==2.0.dev23",
- "bw2data==4.0.dev46",
- "bw2io==0.9.dev31",

these versions are a bit old, what are the problems that impede the use of newer versions of these packages ?
