### 2025-12-23 4d2e72d
improve phrasing in conda installation instructions
