### 2025-12-23 b33ccec
improve phrasing in conda installation instructions for ARM
