### 2025-12-23 2831396
compatibility with newer bw versions
