+++ tests/fixtures/dynamic_biomatrix_db_fixture.py
@@ -116,3 +116,7 @@ def dynamic_biosphere_matrix_db():
             (("bio", "CO2"), 1),
         ]
     )
+    
+    for db in bd.databases:
+        bd.Database(db).register()
+        bd.Database(db).process()
