+++ tests/fixtures/nonunitary_db_fixture.py
@@ -122,3 +122,7 @@ def nonunitary_db():
             (("bio", "CO2"), 1),
         ]
     )
+    
+    for db in bd.databases:
+        bd.Database(db).register()
+        bd.Database(db).process()
