+++ tests/test_temporal_grouping.py
@@ -13,130 +13,92 @@ from bw_temporalis import TemporalDistribution
 
 from bw_timex import TimexLCA
 
+@pytest.mark.usefixtures("temporal_grouping_db_monthly")
+def test_monthly_resolution_score():
 
-# make sure the test db is loaded
-def test_temporal_grouping_db_fixture(temporal_grouping_db):
-    assert len(bd.databases) == 4
+    fu = bd.get_node(database="foreground", code="A")
 
+    database_dates = {
+        "db_2022": datetime.strptime("2022", "%Y"),
+        "db_2024": datetime.strptime("2024", "%Y"),
+        "foreground": "dynamic",
+    }
 
-@pytest.mark.usefixtures("temporal_grouping_db")
-class TestClass_Grouping:
+    tlca = TimexLCA(
+        demand={fu.key: 1},
+        method=("GWP", "example"),
+        database_dates=database_dates,
+    )
+    tlca.build_timeline(
+        temporal_grouping="month",
+        starting_datetime=datetime.strptime("2024-01-02", "%Y-%m-%d"),
+    )
+    tlca.lci()
+    tlca.static_lcia()
+    print(tlca.timeline)
 
-    # yearly resolution is alreday tested in test_electric_vehicle.py
+    expected_score = 1 / 3 * 15 + 1 / 6 * 15 + 1 / 6 * 10 + 1 / 3 * 10
 
-    def test_monthly_resolution_score(self):
+    assert tlca.static_score == expected_score
+    
+@pytest.mark.usefixtures("temporal_grouping_db_daily")
+def test_daily_resolution_score():
 
-        self.fu = bd.get_node(database="foreground", code="A")
+    # first change TD in database to daily (although not strictly necessary)
+    fu = bd.get_node(database="foreground", code="A")
 
-        database_dates = {
-            "db_2022": datetime.strptime("2022", "%Y"),
-            "db_2024": datetime.strptime("2024", "%Y"),
-            "foreground": "dynamic",
-        }
+    database_dates = {
+        "db_2022": datetime.strptime(
+            "2024-01-10", "%Y-%m-%d"
+        ),  # give dbs also a daily resolution (makes calculating the expected score easier)
+        "db_2024": datetime.strptime("2024-01-25", "%Y-%m-%d"),
+        "foreground": "dynamic",
+    }
 
-        self.tlca = TimexLCA(
-            demand={self.fu.key: 1},
-            method=("GWP", "example"),
-            database_dates=database_dates,
-        )
-        self.tlca.build_timeline(
-            temporal_grouping="month",
-            starting_datetime=datetime.strptime("2024-01-02", "%Y-%m-%d"),
-        )
-        self.tlca.lci()
-        self.tlca.static_lcia()
-        print(self.tlca.timeline)
+    tlca = TimexLCA(
+        demand={fu.key: 1},
+        method=("GWP", "example"),
+        database_dates=database_dates,
+    )
+    tlca.build_timeline(
+        temporal_grouping="day",
+        starting_datetime=datetime.strptime("2024-01-25", "%Y-%m-%d"),
+    )
+    tlca.lci()
+    tlca.static_lcia()
+    print(tlca.timeline)
 
-        expected_score = 1 / 3 * 15 + 1 / 6 * 15 + 1 / 6 * 10 + 1 / 3 * 10
+    expected_score = 1 / 3 * 15 + 2 / 9 * 15 + 1 / 9 * 10 + 1 / 3 * 10
 
-        assert self.tlca.static_score == expected_score
+    assert tlca.static_score == pytest.approx(expected_score, rel=0.00005)
 
-    def test_daily_resolution_score(self):
+@pytest.mark.usefixtures("temporal_grouping_db_hourly")
[…]
