+++ tests/fixtures/vehicle_db_fixture.py
@@ -432,7 +432,7 @@ def vehicle_db():
             (("bio", "CO2"), 1),
         ]
     )
-    
+
     for db in bd.databases:
         bd.Database(db).register()
         bd.Database(db).process()
