+++ tests/test_signs.py
@@ -161,7 +161,7 @@ def wastes_db():
             (("biosphere", "CO2"), 1),
         ]
     )
-    
+
     for db in bd.databases:
         bd.Database(db).register()
         bd.Database(db).process()
