### 2025-12-23 b7122cc
add conda forge recipes
