+++ conda-forge-recipes/bw_temporalis/meta.yaml
@@ -1,0 +1,57 @@
+{% set name = "bw_temporalis" %}
+{% set version = "1.2.0" %}
+
+package:
+  name: {{ name|lower }}
+  version: {{ version }}
+
+source:
+  url: https://pypi.io/packages/source/{{ name[0] }}/{{ name }}/{{ name }}-{{ version }}.tar.gz
+  sha256: 7ef666bcc4eff4f03681d82f0b7a57bf6144a4a053da0a0bf2f5ec7e58ff28ab
+
+build:
+  noarch: python
+  number: 0
+  script: {{ PYTHON }} -m pip install . -vv --no-deps --no-build-isolation
+
+requirements:
+  host:
+    - python >=3.9
+    - pip
+    - setuptools
+  run:
+    - python >=3.9
+    - bw2calc >=2.0
+    - bw2data >=4.0
+    - bw_graph_tools
+    - numpy
+    - pandas
+    - scipy
+    - tqdm
+
+test:
+  imports:
+    - bw_temporalis
+  requires:
+    - pip
+  commands:
+    - pip check
+
+about:
+  home: https://github.com/brightway-lca/bw_temporalis
+  summary: Temporal distributions for Brightway LCA framework
+  description: |
+    bw_temporalis is a revision of bw2temporalis for modern Brightway.
+    It enables temporal analysis within the Brightway life cycle assessment
+    framework, handling temporal dynamics in technosphere and biosphere
+    matrices with temporal resolution down to seconds.
+  license: BSD-3-Clause
+  license_family: BSD
+  license_file: LICENSE
+  doc_url: https://github.com/brightway-lca/bw_temporalis
+  dev_url: https://github.com/brightway-lca/bw_temporalis
+
+extra:
+  recipe-maintainers:
+    - cmutel
+    - TimoDiepers
