### 2025-12-24 cc59050
update installation instructions
