### 2026-01-26 e75c9b2
Update dependency bw2calc to v2.3.2
