### 2026-02-18 3f88059
add breadth first graph traversal option
