+++ bw_timex/edge_extractor.py
@@ -1,10 +1,15 @@
+import json
+from collections import deque
 from dataclasses import dataclass
+from datetime import datetime
 from heapq import heappop, heappush
 from numbers import Number
 from typing import Callable
 
 import numpy as np
-from bw_temporalis import TemporalDistribution, TemporalisLCA
+from bw2data.backends.schema import ActivityDataset as AD
+from bw2data.backends.schema import ExchangeDataset as ED
+from bw_temporalis import TemporalDistribution, TemporalisLCA, loader_registry
 
 datetime_type = np.dtype("datetime64[s]")
 timedelta_type = np.dtype("timedelta64[s]")
@@ -245,3 +250,260 @@ class EdgeExtractor(TemporalisLCA):
             raise ValueError(
                 f"Can't join TemporalDistribution and something else: Trying with {type(td_consumer.date)} and {type(td_producer.date)}"
             )
+
+
+class EdgeExtractionBFS:
+    """
+    BFS-based graph traversal for extracting temporal edges from the supply chain.
+
+    Unlike EdgeExtractor (which inherits from TemporalisLCA and uses priority-first
+    traversal with per-subgraph LCA calculations), this class works directly with
+    the technosphere matrix from a bw2calc LCA object and traverses using
+    breadth-first search. This avoids the overhead of computing individual subgraph
+    LCAs for priority ordering.
+
+    Returns the same list[Edge] format as EdgeExtractor, so all downstream code
+    (TimelineBuilder, MatrixModifier, etc.) works unchanged.
+    """
+
+    def __init__(
+        self,
+        lca_object,
+        starting_datetime: datetime | str = "now",
+        edge_filter_function: Callable = None,
+        cutoff: float = 1e-9,
+        static_activity_indices: set[int] | None = None,
+    ) -> None:
+        self.lca_object = lca_object
+        self.edge_ff = edge_filter_function if edge_filter_function else lambda x: False
+        self.cutoff = cutoff
+        self.static_activity_indices = static_activity_indices or set()
+
+        if isinstance(starting_datetime, str):
+            if starting_datetime == "now":
+                starting_datetime = datetime.now()
+            else:
+                starting_datetime = datetime.fromisoformat(starting_datetime)
+
+        self.t0 = TemporalDistribution(
+            np.array([np.datetime64(starting_datetime)]),
+            np.array([1]),
+        )
+
+        self.tech_matrix_csc = self.lca_object.technosphere_matrix.tocsc()
+        self._ad_cache = {}
+
+    def _get_activity_dataset(self, activity_id: int) -> AD:
+        if activity_id not in self._ad_cache:
+            self._ad_cache[activity_id] = AD.get(AD.id == activity_id)
+        return self._ad_cache[activity_id]
+
+    def _get_exchange(self, input_id: int, output_id: int):
+        """Look up exchange between two activities. Returns ExchangeDataset or None."""
+        inp = self._get_activity_dataset(input_id)
+        outp = self._get_activity_dataset(output_id)
+        exchanges = list(
+            ED.select().where(
+                ED.input_code == inp.code,
+                ED.input_database == inp.database,
+                ED.output_code == outp.code,
+                ED.output_database == outp.database,
+            )
+        )
+        if len(exchanges) == 1:
+            return exchanges[0]
+        elif len(exchanges) > 1:
+            raise ValueError(
+                f"Found {len(exchanges)} exchanges between {input_id} and {output_id}"
+            )
+        return None
+
+    def _get_exchange_td_and_type(self, input_id: int, output_id: int):
+        """
+        Get temporal distribution and edge type for an exchange.
+
+        Returns (td_or_amount, edge_type) where td_or_amount is either a
+        TemporalDistribution or a float (the signed matrix value).
+        """
+        exchange = self._get_exchange(input_id, output_id)
+
[…]
