### 2026-02-19 4a292d4
add bg graph traversal
