### 2026-02-19 7359f61
improve bfs docstrings and rename EdgeExtractorBFS
