### 2026-02-23 1417c28
feat: foreground evolution
