+++ tests/test_temporal_evolution.py
@@ -1,0 +1,326 @@
+"""Tests for temporal evolution of foreground exchange amounts."""
+
+from datetime import datetime
+
+import pytest
+
+from bw_timex.utils import get_temporal_evolution_factor
+
+
+class TestGetTemporalEvolutionFactor:
+    """Tests for the get_temporal_evolution_factor utility function."""
+
+    def test_exact_match(self):
+        evolution = {
+            datetime(2020, 1, 1): 1.0,
+            datetime(2030, 1, 1): 0.75,
+        }
+        assert get_temporal_evolution_factor(evolution, datetime(2020, 1, 1)) == 1.0
+        assert get_temporal_evolution_factor(evolution, datetime(2030, 1, 1)) == 0.75
+
+    def test_linear_interpolation(self):
+        evolution = {
+            datetime(2020, 1, 1): 1.0,
+            datetime(2030, 1, 1): 0.5,
+        }
+        result = get_temporal_evolution_factor(evolution, datetime(2025, 1, 1))
+        assert result == pytest.approx(0.75, abs=0.01)
+
+    def test_clamp_below(self):
+        evolution = {
+            datetime(2020, 1, 1): 1.0,
+            datetime(2030, 1, 1): 0.5,
+        }
+        assert get_temporal_evolution_factor(evolution, datetime(2010, 1, 1)) == 1.0
+
+    def test_clamp_above(self):
+        evolution = {
+            datetime(2020, 1, 1): 1.0,
+            datetime(2030, 1, 1): 0.5,
+        }
+        assert get_temporal_evolution_factor(evolution, datetime(2040, 1, 1)) == 0.5
+
+    def test_single_point(self):
+        evolution = {datetime(2020, 1, 1): 0.8}
+        assert get_temporal_evolution_factor(evolution, datetime(2025, 1, 1)) == 0.8
+
+    def test_three_points(self):
+        evolution = {
+            datetime(2020, 1, 1): 1.0,
+            datetime(2030, 1, 1): 0.5,
+            datetime(2040, 1, 1): 0.2,
+        }
+        # Midpoint between 2030 and 2040
+        result = get_temporal_evolution_factor(evolution, datetime(2035, 1, 1))
+        assert result == pytest.approx(0.35, abs=0.01)
+
+    def test_empty_dict_returns_one(self):
+        assert get_temporal_evolution_factor({}, datetime(2025, 1, 1)) == 1.0
+
+    def test_none_returns_one(self):
+        assert get_temporal_evolution_factor(None, datetime(2025, 1, 1)) == 1.0
+
+
+import bw2data as bd
+import numpy as np
+from bw2data.tests import bw2test
+from bw_temporalis import TemporalDistribution
+
+from bw_timex import TimexLCA
+
+
+@pytest.mark.usefixtures("temporal_evolution_db")
+class TestTemporalEvolutionFactors:
+    """Integration tests for temporal_evolution_factors on exchanges."""
+
+    @pytest.fixture(autouse=True)
+    def setup_method(self, temporal_evolution_db):
+        self.consumer = bd.get_node(database="foreground", code="consumer")
+        database_dates = {
+            "db_2020": datetime.strptime("2020", "%Y"),
+            "db_2030": datetime.strptime("2030", "%Y"),
+            "foreground": "dynamic",
+        }
+        self.tlca = TimexLCA(
+            demand={self.consumer.key: 1},
+            method=("GWP", "example"),
+            database_dates=database_dates,
+        )
+
+    def test_timeline_has_temporal_evolution_column(self):
+        self.tlca.build_timeline(
+            starting_datetime=datetime.strptime("2024-01-01", "%Y-%m-%d")
+        )
[…]
