+++ bw_timex/__init__.py
@@ -4,5 +4,10 @@ from .helper_classes import SetList
 from .matrix_modifier import MatrixModifier
 from .timeline_builder import TimelineBuilder
 from .timex_lca import TimexLCA
+from .utils import (
+    add_temporal_distribution_to_exchange,
+    add_temporal_evolution_to_exchange,
+    get_temporal_evolution_factor,
+)
 
 __version__ = "0.3.4"
