### 2026-02-24 #9621 @arfon
@rahuldevikar @mahajanhrishikesh – 👋 would you be willing to review this submission for JOSS? The submission under consideration is _bw_timex: A Python Package for Time-Explicit Life Cycle Assessment_

Many thanks
Arfon
