### 2026-02-24 c2c6945
Add demand validation in TimexLCA.__init__ (fixes #160)
