### 2026-02-24 20925fa
add feature of temporal evolution to theory and landing page
