### 2026-02-24 56aa1d4
Clarify temporal distribution description in documentation
Updated terminology from 'time-explicit databases' to 'time-specific background databases' for clarity.
