### 2026-02-24 aa99fa5
use uv for testing
