+++ .github/workflows/python-package-deploy.yml
@@ -16,22 +16,12 @@ jobs:
       id-token: write
     steps:
     - uses: actions/checkout@main
+    - name: Install uv
+      uses: astral-sh/setup-uv@v5
     - name: Set up Python 3.11
-      uses: actions/setup-python@v6
-      with:
-        python-version: "3.11"
-    - name: Install pypa/build
-      run: >-
-        python -m
-        pip install
-        build
-        --user
+      run: uv python install 3.11
     - name: Build a binary wheel and a source tarball
-      run: >-
-        python -m
-        build
-        --outdir dist/
-        .
+      run: uv build --out-dir dist/
     - name: Publish distribution 📦 to Test PyPI
       uses: pypa/gh-action-pypi-publish@release/v1
       with:
