### 2026-02-24 5054570
chore(deps): update astral-sh/setup-uv action to v7
