+++ .github/workflows/python-package-deploy.yml
@@ -17,7 +17,7 @@ jobs:
     steps:
     - uses: actions/checkout@main
     - name: Install uv
-      uses: astral-sh/setup-uv@v5
+      uses: astral-sh/setup-uv@v7
     - name: Set up Python 3.11
       run: uv python install 3.11
     - name: Build a binary wheel and a source tarball
