+++ .github/workflows/python-test.yml
@@ -42,7 +42,7 @@ jobs:
       - uses: actions/checkout@v6
 
       - name: Install uv
-        uses: astral-sh/setup-uv@v5
+        uses: astral-sh/setup-uv@v7
 
       - name: Set up Python ${{ matrix.py-version }}
         run: uv python install ${{ matrix.py-version }}
