### 2026-02-25 #9621 @rahuldevikar
# JOSS Review Checklist — `bw_timex` (#9621)

**Checklist version:** [pre-2026-submission](https://joss.readthedocs.io/en/pre-2026-submission/review_checklist.html)

---

## Conflict of Interest

- [x] I confirm that I have read the [JOSS conflict of interest policy](https://joss.readthedocs.io/en/pre-2026-submission/reviewer_guidelines.html#joss-conflict-of-interest-policy) and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.

- [x] **Repository:** Is the source code for this software available at the [repository url](https://github.com/brightway-lca/bw_timex)?
- [x] **Contribution and authorship:** Has the submitting author made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
