### 2026-02-25 brightway-lca/bw_timex#170 @rahuldevikar
Add test coverage reporting (e.g., Codecov) and README badge
Hi, this is a suggestion as part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/9621).

I noticed that pytest-cov is already included in the testing extras in pyproject.toml, which is great. However, there doesn't appear to be a coverage report published or a coverage badge displayed in the README.

Adding a coverage reporting integration (e.g., [Codecov](https://codecov.io/) or [Coveralls](https://coveralls.io/)) to the GitHub Actions CI workflow would:

- Provide transparency about the extent of test coverage
- Make it easy for contributors to see if their changes affect coverage
- Add a visible coverage badge to the README alongside the existing badges

This is a non-blocking suggestion — the existing test suite and CI are solid.
