### 2026-02-25 brightway-lca/bw_timex#171 @rahuldevikar
Enable cross-platform CI testing (Windows, macOS)
Hi, this is a suggestion as part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/9621).

The current GitHub Actions CI workflow (.github/workflows/python-test.yml) runs only on ubuntu-latest. I noticed there's a commented-out line for windows-latest and macos-latest:

```yaml
os: [ubuntu-latest] # , windows-latest, macos-latest
```

Since the installation documentation addresses multiple platforms and the package is pure Python, it would be helpful to enable CI on Windows and macOS as well. This would:

- Catch any platform-specific issues (e.g., path handling, file encoding)
- Provide additional confidence for users on those platforms
- Be a straightforward change given the existing configuration

This is a non-blocking suggestion — the package is likely cross-platform compatible already given its pure-Python nature.
