### 2026-02-25 brightway-lca/bw_timex#172 @rahuldevikar
Add a CHANGELOG.md to the repository root
Hi, this is a suggestion as part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/9621).

The project has excellent release notes on the [GitHub Releases page](https://github.com/brightway-lca/bw_timex/releases), covering all 24+ releases with clear descriptions of changes. However, there is no `CHANGELOG.md` file in the repository root.

Adding a `CHANGELOG.md` would:

- Improve discoverability of version history for users who access the package through PyPI or conda rather than GitHub
- Follow common open-source conventions (e.g., [Keep a Changelog](https://keepachangelog.com/))
- Complement the existing GitHub Releases with an in-repo record

This is a non-blocking suggestion.
