### 2026-02-25 brightway-lca/bw_timex#172 @TimoDiepers
Hi @rahuldevikar, our changes are tracked in the file `CHANGES.md` in the repo root, according to Keep a Changelog (with v1.0.0 planned to be released with the JOSS publication) . We've carried over this naming instead of `CHANGELOG.md` to be consistent with the convention in all other brightway packages.
