+++ .github/workflows/python-test.yml
@@ -51,4 +51,11 @@ jobs:
         run: uv sync --extra testing
 
       - name: Test with pytest
-        run: uv run pytest
+        run: uv run pytest --cov-report xml
+
+      - name: Upload coverage to Codecov
+        if: matrix.os == 'ubuntu-latest' && matrix.py-version == '3.13'
+        uses: codecov/codecov-action@v5
+        with:
+          file: ./coverage.xml
+          fail_ci_if_error: false
