+++ codecov.yml
@@ -1,11 +1,0 @@
-codecov:
-  branch: main
-
-coverage:
-  status:
-    project:
-      default:
-        target: auto
-    patch:
-      default:
-        target: auto
