### 2026-02-25 d034920
Add JOSS badge
