### 2026-02-25 #9621 @rahuldevikar
## Review Summary

Thanks for the opportunity to review `bw_timex`. I've gone through the repository, documentation, paper, and test suite in detail. Overall, this is a strong submission — the package addresses a genuine methodological gap in the LCA tooling landscape and is backed by solid engineering practices and excellent documentation.

### Strengths

- **Clear scholarly contribution:** `bw_timex` fills a well-articulated gap by unifying temporal distribution (dLCA) and temporal evolution (pLCA) in a single framework. The authors make a compelling case that this combination is not well-served by existing tools, and the approach is supported by a peer-reviewed methodology paper (Müller, Diepers et al., 2025, [doi:10.1007/s11367-025-02539-3](https://doi.org/10.1007/s11367-025-02539-3)).
- **Excellent documentation:** The docs are exemplary — a comprehensive theory section with worked examples, a step-by-step getting-started guide, an interactive decision tree, three real-world example notebooks (including one that reproduces the published paper's results), and a full API reference.
- **Solid test suite and CI:** 9 test files covering core functionality, hooked up to GitHub Actions CI across Python 3.11, 3.12, and 3.13.
- **Good community practices:** CONTRIBUTING.md, CODE_OF_CONDUCT.md, GitHub Discussions, and direct maintainer contact information are all in place.

### Suggestions (Non-Blocking)

I've filed the following issues on the repository:

1. **Add CI coverage reporting and badge** — `pytest-cov` is configured but no coverage report or badge is published. See [brightway-lca/bw_timex#170](https://github.com/brightway-lca/bw_timex/issues)
2. **Enable cross-platform CI (Windows, macOS)** — The CI config has these commented out; enabling them would strengthen cross-platform confidence. See [brightway-lca/bw_timex#171](https://github.com/brightway-lca/bw_timex/issues)
3. **Add a CHANGELOG.md** — Release notes on GitHub Releases are excellent, but an in-repo changelog would improve discoverability for users accessing the package through PyPI or conda. See [brightway-lca/bw_timex#172](https://github.com/brightway-lca/bw_timex/issues)

Additionally, three suggestions for the review thread:

4. **Version update:** The submitted version is v0.3.0, but v0.3.4 (Jan 5, 2026) includes important Brightway compatibility fixes. I'd recommend updating the submitted version before acceptance so the archived DOI reflects these improvements.
5. **State of the field labeling:** The comparison with existing tools (Temporalis, premise, Futura, pathways) is well-done, but it's embedded within the "Statement of need" rather than labeled as its own subsection. Adding a brief `### State of the field` heading in the paper would more explicitly satisfy that checklist item.
6. **Consider citing Levasseur et al. (2010):** Since dynamic characterization is highlighted as a key capability of `bw_timex`, consider adding a reference to the foundational work on time-dependent characterization factors — Levasseur et al. (2010), "Considering time in LCA: dynamic LCA and its application to global warming impact assessments" ([doi:10.1021/es9030003](https://doi.org/10.1021/es9030003)). This paper is already referenced in the `bw_timex` documentation's theory section when discussing fixed vs. flexible time horizons.

### Checklist

All items on the pre-2026 JOSS review checklist are met. Please see my completed checklist in the comment above.

I'm happy to discuss any of the above, and looking forward to the authors' response. Nice work on the package!
