### 2026-02-25 brightway-lca/bw_timex#172 @rahuldevikar
Make sense. Thanks
