+++ joss/paper.md
@@ -53,7 +53,7 @@ LCA traditionally assumes a static system, where all processes occur simultaneou
 
 # Workflow
 
-A time-explicit LCA with `bw_timex` follows four main steps, as illustrated in \autoref{fig:workflow}. First, a conventional product system model is temporalized by adding process-relative temporal distributions (rTDs) to the exchanges (cf. @Cardellini:2018). These rTDs describe how the amount of a technosphere or biosphere exchange is distributed over time, relative to the consuming or emitting process. In Step 2, a timeline of technosphere exchanges is constructed by convolving rTDs along the supply chain, starting from the absolute reference time for the demand, which is defined by the user. In Step 3, the exchanges in the timeline are re-linked to time-specific background databases that reflect the technology landscape at specific points in time. Based on the temporally re-linked product system, a time-explicit LCI is calculated, preserving the timing of processes and emissions. The inventory is calculated following the conventional matrix-based LCA formulation [@Heijungs:2002], with the time dimension embedded in the matrices through additional row/column pairs. In Step 4, these emissions are characterized, either using standard characterization factors or by applying dynamic characterization functions that take the emissions’ timing into account.
+A time-explicit LCA with `bw_timex` follows four main steps, as illustrated in \autoref{fig:workflow}. First, a conventional product system model is temporalized by adding process-relative temporal distributions (rTDs) to the exchanges (cf. [@Cardellini:2018]). These rTDs describe how the amount of a technosphere or biosphere exchange is distributed over time, relative to the consuming or emitting process. In Step 2, a timeline of technosphere exchanges is constructed by convolving rTDs along the supply chain, starting from the absolute reference time for the demand, which is defined by the user. In Step 3, the exchanges in the timeline are re-linked to time-specific background databases that reflect the technology landscape at specific points in time. Based on the temporally re-linked product system, a time-explicit LCI is calculated, preserving the timing of processes and emissions. The inventory is calculated following the conventional matrix-based LCA formulation [@Heijungs:2002], with the time dimension embedded in the matrices through additional row/column pairs. In Step 4, these emissions are characterized, either using standard characterization factors or by applying dynamic characterization functions that take the emissions’ timing into account.
 
 ![Workflow of a time-explicit LCA with `bw_timex`.\label{fig:workflow}](workflow.pdf){width=9cm}
 
