+++ joss/paper.bib
@@ -140,3 +140,23 @@
 	month = oct,
 	year = {2025},
 }
+
+@software{Sacchi:2026,
+title = {TRAILS: Temporal Routing and Aggregation of Impacts across Life-cycle Systems},
+author = { Sacchi , R.},
+repository-code = {https://github.com/Laboratory-for-Energy-Systems-Analysis/trails},
+url = {https://trails.readthedocs.io/en/latest/},
+version = {v1.0.0},
+year = {2026}
+}
+
+@article{LangQuantzendorff:2025,
+  title = {Prosperdyn---a Tool to Describe Dynamic Transitions in Prospective Life Cycle Assessment},
+  author = {{Lang-Quantzendorff}, L. and Beermann, M.},
+  year = 2025,
+  month = oct,
+  journal = {The International Journal of Life Cycle Assessment},
+  issn = {0948-3349, 1614-7502},
+  doi = {10.1007/s11367-025-02515-x},
+  urldate = {2025-10-22},
+}
