+++ joss/paper.md
@@ -37,29 +37,35 @@ bibliography: paper.bib
 
 # Summary
 
-`bw_timex` is a Python package for time-explicit Life Cycle Assessment (LCA). Unlike conventional LCA, time-explicit LCA allows the quantification of environmental impacts of products and processes *over time*, considering their temporal distribution and evolution. As such, `bw_timex` allows to account simultaneously for:
+`bw_timex` is a Python package for time-explicit Life Cycle Assessment (LCA). Unlike conventional LCA, time-explicit LCA allows the quantification of environmental impacts of products and processes *over time*, considering their temporal distribution and evolution. As such, `bw_timex` enables simultaneously accounting for:
 
 - the timing of processes throughout the supply chain (e.g., end-of-life treatment occurs 20 years after production),
-- variable and/or evolving supply chains and technologies (e.g., increasing shares of renewable electricity in the future), and
+- variable and/or evolving supply chains and technologies (e.g., increasing shares of renewable electricity or higher process efficiencies in the future), and
 - the timing of emissions (enabling dynamic characterization).
 
 To achieve this, `bw_timex` uses graph traversal to convolve process-relative temporal distributions through the supply chain. From the resulting timeline of technosphere exchanges, Life Cycle Inventories (LCIs) are automatically linked across time-specific background databases. The resulting time-explicit LCI reflects the current technology status within the product system at the actual time of each process. Moreover, `bw_timex` preserves the timing of emissions, enabling both dynamic and static Life Cycle Impact Assessment.
 
 # Statement of need
 
-LCA traditionally assumes a static system, where all processes occur simultaneously and do not change over time [@Heijungs:2002]. To add a temporal dimension in LCA, the fields of dynamic LCA (dLCA) and prospective LCA (pLCA) have emerged. While dLCA focuses on when processes and emissions occur and how impacts are distributed over time (*temporal distribution*), it typically assumes that the underlying product system remains the same [@Beloin:2020]. Conversely, while pLCA tracks how processes evolve (*temporal evolution*) using future scenarios, it generally only assesses a single (future) point in time, ignoring that processes occur at different times across a product’s life cycle [@Arvidsson:2024]. Both fields have seen open-source tool development, including `Temporalis` [@Cardellini:2018] for dLCA and `premise` [@Sacchi:2022], `Futura` [@Joyce:2022] and `pathways` [@Sacchi:2024] for pLCA. However, a comprehensive open-source package that supports consideration of both temporal distribution and evolution is currently lacking.
+LCA traditionally assumes a static system, where all processes occur simultaneously and do not change over time [@Heijungs:2002]. To add a temporal dimension to LCA, the fields of dynamic LCA (dLCA) and prospective LCA (pLCA) have emerged. While dLCA focuses on when processes and emissions occur and how impacts are distributed over time (*temporal distribution*), it typically assumes that the underlying product system remains the same [@Beloin:2020]. Conversely, while pLCA tracks how processes evolve (*temporal evolution*) using future scenarios, it generally only assesses a single (future) point in time, ignoring that processes occur at different times across a product’s life cycle [@Arvidsson:2024].
 
-`bw_timex` addresses this gap by providing a framework for time-explicit LCA calculations within the `Brightway` ecosystem [@Mutel:2017]. It combines considerations of temporal distribution and evolution by accounting for both the timing of processes and emissions as well as the state of the product system at the respective points in time. This makes `bw_timex` particularly useful for studies involving variable or strongly evolving product systems, long-lived products, biogenic carbon and scenario analyses.
[…]
