### 2026-02-26 #9621 @TimoDiepers
Hi @rahuldevikar, thanks for your in-depth review. For points 1-3, the corresponding issues have been addressed and closed.

About 4 (Version update): Indeed, we've added some new features since the original submission a while back. We intend to make a new release after all reviews have been addressed, which we will then also archive upon acceptance.

About 5 (State of the field labelling): Thanks for pointing this out. We fully agree and have added an explicit State of the field section, also incorporating some new tools that have been published since our JOSS submission. Also, we've made small updates to the "Workflow" section of the paper to include the new features that have been added since original submission.

About 6 (Citing Levasseur et al.): In our view, dynamic characterization as described by Levasseur et al. is not the core functionality of bw_timex. For its dynamic characterization capabilities, bw_timex depends on the separate library [dynamic_characterization](https://github.com/brightway-lca/dynamic_characterization), which is why we avoided going into too much detail on this topic in the present paper. An in-detail description of linking this and other dynamic characterization methods can be found in Müller et al. 2025, which we refer to in our "Further reading" section.
