### 2026-02-26 #9621 @rahuldevikar
Thanks @TimoDiepers for the quick turnaround on all of this.

Glad to see the following issues addressed. The approach for point 4 (new release upon acceptance) makes sense.
- [brightway-lca/bw_timex#170](https://github.com/brightway-lca/bw_timex/issues)
- [brightway-lca/bw_timex#171](https://github.com/brightway-lca/bw_timex/issues)
- [brightway-lca/bw_timex#172](https://github.com/brightway-lca/bw_timex/issues)

Regarding point 5, thank you for going through the suggestion — adding the explicit State of the field section and updating the Workflow for new features strengthens the paper.

On point 6, that's a fair point — The reference to Müller et al. 2025 for the broader context should be good enough since dynamic characterization is handled by the separate library.

I've reviewed the updated PDF and all my checklist items remain satisfied. Happy with where things stand from my end.

-----------------------------------------------------------------------------------------------------------------------------------------------
@arfon - could you add me as an official reviewer? I've already posted my checklist and review summary above. The bot blocked my `generate my checklist` command since I'm not listed yet.
