### 2026-03-05 6a4534c
fix(deps): update dependency bw2data to v4.6
