+++ pyproject.toml
@@ -31,7 +31,7 @@ classifiers = [
 requires-python = ">=3.11"
 dependencies = [
     "bw2calc==2.4.0",
-    "bw2data==4.5.4",
+    "bw2data==4.6",
     "bw_graph_tools>0.4",
     "numpy",
     "pandas",
