### 2026-03-05 #9621 @arfon
@mahajanhrishikesh and @rahuldevikar – thanks so much for your quick responses. You should be able to ask Editorialbot to create your review checklists now.
