### 2026-03-22 #9621 @mahajanhrishikesh
Hey team, I have done my first round for the review. Most of the work seems to be in order except for these two issues:

1. https://github.com/brightway-lca/bw_timex/issues/178 an import issue ... probably due to a recent change in bw data.
2. I am not sure if I am supposed to run the benchmarking notebook myself to independently verify, it seems like I should be doing that given that the paper does the claim of the performance being better. To run the notebook however I do not have access to the .tar.gz file it references inside it. If it is not required we can close that out.

Rest of it looks great and is an amazingly detailed effort, great job guys!
