### 2026-03-22 #9621 @TimoDiepers
Hello @mahajanhrishikesh, thanks for your review. About point 2, we unfortunately can't publish the database file because of ecoinvent licensing issues. The comparison in there is not meant as a performance claim anyway, rather just to have some baseline to compare potential future performance improvements to. If you have access to an ecoinvent license, I'm happy to provide the database file to you if you still want to check for yourself.
