### 2026-03-23 1182285
test: add additional tests of non-core functionality
