+++ tests/test_utils.py
@@ -1,0 +1,369 @@
+"""Tests for utility functions in bw_timex.utils."""
+
+from datetime import datetime
+
+import bw2data as bd
+import numpy as np
+import pandas as pd
+import pytest
+from bw2data.errors import MultipleResults, UnknownObject
+from bw_temporalis import TemporalDistribution
+
+from bw_timex.utils import (
+    add_flows_to_characterization_functions,
+    add_temporal_distribution_to_exchange,
+    add_temporal_evolution_to_exchange,
+    convert_date_string_to_datetime,
+    extract_date_as_integer,
+    extract_date_as_string,
+    get_exchange,
+    get_temporal_evolution_factor,
+    resolve_temporalized_node_name,
+    round_datetime,
+)
+
+# --- extract_date_as_integer ---
+
+
+class TestExtractDateAsInteger:
+    def test_year(self):
+        assert extract_date_as_integer(datetime(2023, 3, 15), "year") == 2023
+
+    def test_month(self):
+        assert extract_date_as_integer(datetime(2023, 3, 15), "month") == 202303
+
+    def test_day(self):
+        assert extract_date_as_integer(datetime(2023, 3, 15), "day") == 20230315
+
+    def test_hour(self):
+        assert extract_date_as_integer(datetime(2023, 3, 15, 14), "hour") == 2023031514
+
+    def test_default_is_year(self):
+        assert extract_date_as_integer(datetime(2023, 7, 1)) == 2023
+
+    def test_invalid_resolution(self):
+        with pytest.raises(ValueError, match="Invalid time_res"):
+            extract_date_as_integer(datetime(2023, 1, 1), "minute")
+
+
+# --- extract_date_as_string ---
+
+
+class TestExtractDateAsString:
+    def test_year(self):
+        assert extract_date_as_string(datetime(2023, 3, 15), "year") == "2023"
+
+    def test_month(self):
+        assert extract_date_as_string(datetime(2023, 3, 15), "month") == "202303"
+
+    def test_day(self):
+        assert extract_date_as_string(datetime(2023, 3, 15), "day") == "20230315"
+
+    def test_hour(self):
+        assert extract_date_as_string(datetime(2023, 3, 15, 14), "hour") == "2023031514"
+
+    def test_invalid_grouping(self):
+        with pytest.raises(ValueError, match="not a valid option"):
+            extract_date_as_string(datetime(2023, 1, 1), "minute")
+
+
+# --- convert_date_string_to_datetime ---
+
+
+class TestConvertDateStringToDatetime:
+    def test_year(self):
+        assert convert_date_string_to_datetime("year", "2023") == datetime(2023, 1, 1)
+
+    def test_month(self):
+        assert convert_date_string_to_datetime("month", "202303") == datetime(
+            2023, 3, 1
+        )
+
+    def test_day(self):
+        assert convert_date_string_to_datetime("day", "20230315") == datetime(
+            2023, 3, 15
+        )
+
+    def test_hour(self):
+        assert convert_date_string_to_datetime("hour", "2023031514") == datetime(
+            2023, 3, 15, 14
+        )
+
+    def test_invalid_grouping(self):
+        with pytest.raises(ValueError, match="not a valid option"):
+            convert_date_string_to_datetime("minute", "2023")
+
+
+# --- round_datetime ---
+
+
+class TestRoundDatetime:
+    # Year rounding
+    def test_round_year_down(self):
+        result = round_datetime(datetime(2023, 3, 15), "year")
+        assert result == pd.Timestamp("2023-01-01")
+
+    def test_round_year_up(self):
+        result = round_datetime(datetime(2023, 9, 15), "year")
+        assert result == pd.Timestamp("2024-01-01")
+
+    def test_round_year_boundary(self):
[…]
