+++ tests/test_timex_lca.py
@@ -1,0 +1,421 @@
+"""Tests for TimexLCA methods: properties, labelling, error paths, and edge cases."""
+
+from datetime import datetime
+
+import bw2data as bd
+import numpy as np
+import pandas as pd
+import pytest
+from dynamic_characterization.classes import CharacterizedRow
+
+from bw_timex import TimexLCA
+
+# ─── Tests for score properties and error paths ───
+
+
+@pytest.mark.usefixtures("temporal_grouping_db_monthly")
+class TestTimexLCAProperties:
+
+    @pytest.fixture(autouse=True)
+    def setup_tlca(self, temporal_grouping_db_monthly):
+        fu = bd.get_node(database="foreground", code="A")
+        database_dates = {
+            "db_2022": datetime.strptime("2022", "%Y"),
+            "db_2024": datetime.strptime("2024", "%Y"),
+            "foreground": "dynamic",
+        }
+        self.tlca = TimexLCA(
+            demand={fu.key: 1},
+            method=("GWP", "example"),
+            database_dates=database_dates,
+        )
+
+    def test_base_score(self):
+        score = self.tlca.base_score
+        assert isinstance(score, float)
+        assert score != 0
+
+    def test_static_score_before_lci_raises(self):
+        with pytest.raises(AttributeError, match="LCI not yet calculated"):
+            _ = self.tlca.static_score
+
+    def test_dynamic_score_before_dynamic_lcia_raises(self):
+        with pytest.raises(AttributeError, match="Characterized inventory not yet"):
+            _ = self.tlca.dynamic_score
+
+    def test_static_lcia_before_lci_raises(self):
+        with pytest.raises(AttributeError, match="LCI not yet calculated"):
+            self.tlca.static_lcia()
+
+    def test_lci_before_timeline_raises(self):
+        with pytest.raises(AttributeError, match="Timeline not yet built"):
+            self.tlca.lci()
+
+    def test_lci_no_expand_no_dynamic_raises(self):
+        self.tlca.build_timeline(
+            starting_datetime=datetime.strptime("2024-01-02", "%Y-%m-%d"),
+        )
+        with pytest.raises(ValueError, match="not possible to skip"):
+            self.tlca.lci(expand_technosphere=False, build_dynamic_biosphere=False)
+
+
+# ─── Tests for labelling / dataframe methods ───
+
+
+@pytest.mark.usefixtures("temporal_grouping_db_monthly")
+class TestTimexLCALabellingMethods:
+
+    @pytest.fixture(autouse=True)
+    def setup_tlca(self, temporal_grouping_db_monthly):
+        fu = bd.get_node(database="foreground", code="A")
+        database_dates = {
+            "db_2022": datetime.strptime("2022", "%Y"),
+            "db_2024": datetime.strptime("2024", "%Y"),
+            "foreground": "dynamic",
+        }
+        self.tlca = TimexLCA(
+            demand={fu.key: 1},
+            method=("GWP", "example"),
+            database_dates=database_dates,
+        )
+        self.tlca.build_timeline(
+            starting_datetime=datetime.strptime("2024-01-02", "%Y-%m-%d"),
+        )
+        self.tlca.lci(expand_technosphere=True, build_dynamic_biosphere=True)
+        self.tlca.static_lcia()
+
+    def test_create_labelled_technosphere_dataframe(self):
+        df = self.tlca.create_labelled_technosphere_dataframe()
+        assert isinstance(df, pd.DataFrame)
+        assert df.shape[0] > 0
+        assert df.shape[0] == df.shape[1]  # square matrix
+
+    def test_create_labelled_biosphere_dataframe(self):
+        df = self.tlca.create_labelled_biosphere_dataframe()
+        assert isinstance(df, pd.DataFrame)
+        assert df.shape[0] > 0
+
[…]
