### 2026-03-23 be8c0e8
validate inputs with pydantic
