+++ bw_timex/validation.py
@@ -1,0 +1,215 @@
+from datetime import datetime
+from typing import Callable, Literal, Optional, Union
+
+import bw2data as bd
+from bw_temporalis import TemporalDistribution
+from pydantic import BaseModel, Field, field_validator, model_validator
+
+
+class TimexLCAInputs(BaseModel):
+    """Validates inputs to TimexLCA.__init__"""
+
+    model_config = {"arbitrary_types_allowed": True}
+
+    demand: dict
+    method: tuple
+    database_dates: Optional[dict] = None
+
+    @field_validator("demand")
+    @classmethod
+    def validate_demand(cls, v: dict) -> dict:
+        if not v:
+            raise ValueError("demand must be a non-empty dictionary.")
+        for key, value in v.items():
+            if not isinstance(value, (int, float)):
+                raise ValueError(
+                    f"demand values must be numeric, got {type(value).__name__} for key {key}."
+                )
+        return v
+
+    @field_validator("method")
+    @classmethod
+    def validate_method(cls, v: tuple) -> tuple:
+        if not v:
+            raise ValueError("method must be a non-empty tuple.")
+        if not all(isinstance(item, str) for item in v):
+            raise ValueError(
+                f"method must be a tuple of strings, got types: {[type(item).__name__ for item in v]}."
+            )
+        if v not in bd.methods:
+            raise ValueError(
+                f"Method {v} not found in the current Brightway project. "
+                f"Available methods can be listed with `bw2data.methods`."
+            )
+        return v
+
+    @field_validator("database_dates")
+    @classmethod
+    def validate_database_dates(cls, v: Optional[dict]) -> Optional[dict]:
+        if v is None:
+            return v
+        if not v:
+            raise ValueError(
+                "database_dates must be a non-empty dictionary if provided."
+            )
+        for database, value in v.items():
+            if not isinstance(database, str):
+                raise ValueError(
+                    f"database_dates keys must be strings (database names), got {type(database).__name__}."
+                )
+            if not (value == "dynamic" or isinstance(value, datetime)):
+                raise ValueError(
+                    f"The value {value} for database '{database}' is neither 'dynamic' nor a datetime object. Check the format of the database_dates."
+                )
+            if database not in bd.databases:
+                raise ValueError(
+                    f"Database '{database}' not available in the Brightway2 project."
+                )
+        return v
+
+
+class BuildTimelineInputs(BaseModel):
+    """Validates inputs to TimexLCA.build_timeline"""
+
+    model_config = {"arbitrary_types_allowed": True}
+
+    starting_datetime: Union[datetime, str] = "now"
+    temporal_grouping: Literal["year", "month", "day", "hour"] = "year"
+    interpolation_type: Literal["linear", "closest", "nearest"] = "linear"
+    edge_filter_function: Optional[Callable] = None
+    cutoff: float = Field(default=1e-9, gt=0)
+    max_calc: int = Field(default=2000, gt=0)
+    graph_traversal: Literal["priority", "bfs"] = "priority"
+
+    @field_validator("starting_datetime")
+    @classmethod
+    def validate_starting_datetime(
+        cls, v: Union[datetime, str]
+    ) -> Union[datetime, str]:
+        if isinstance(v, str):
+            if v == "now":
+                return v
+            # Try to parse the string as a date
+            try:
+                datetime.fromisoformat(v)
+            except ValueError:
+                raise ValueError(
+                    f"starting_datetime string must be 'now' or a valid ISO format date string, got '{v}'."
+                )
+        return v
+
+    @field_validator("interpolation_type")
+    @classmethod
[…]
