+++ bw_timex/utils.py
@@ -469,6 +469,9 @@ def add_temporal_distribution_to_exchange(
     None
         The exchange is saved with the temporal distribution.
     """
+    from .validation import TemporalDistributionExchangeInputs
+
+    TemporalDistributionExchangeInputs(temporal_distribution=temporal_distribution)
     exchange = get_exchange(**kwargs)
     exchange["temporal_distribution"] = temporal_distribution
     exchange.save()
@@ -496,14 +499,12 @@ def add_temporal_evolution_to_exchange(
     None
         The exchange is saved with the temporal evolution data.
     """
-    if (
-        temporal_evolution_factors is not None
-        and temporal_evolution_amounts is not None
-    ):
-        raise ValueError(
-            "'temporal_evolution_factors' and 'temporal_evolution_amounts' are "
-            "mutually exclusive — use one or the other."
-        )
+    from .validation import TemporalEvolutionExchangeInputs
+
+    TemporalEvolutionExchangeInputs(
+        temporal_evolution_factors=temporal_evolution_factors,
+        temporal_evolution_amounts=temporal_evolution_amounts,
+    )
     exchange = get_exchange(**kwargs)
     if temporal_evolution_factors is not None:
         exchange["temporal_evolution_factors"] = temporal_evolution_factors
