### 2026-03-23 ab938dd
move old demand validation to pydantic
