+++ joss/paper.md
@@ -53,13 +53,13 @@ LCA traditionally assumes a static system, where all processes occur simultaneou
 
 # State of the field
 
-Existing dLCA tools such as `Temporalis` [@Cardellini:2018] handle temporal distributions but not system evolution. Conversely, pLCA tools like `premise` [@Sacchi:2022], `Futura` [@Joyce:2022], and `pathways` [@Sacchi:2024] model evolving systems but not temporal distributions within the supply chain. Two recent tools combine both temporal distribution and evolution: `ProsperDyn` [@LangQuantzendorff:2025] and `TRAILS` [@Sacchi:2026]. `ProsperDyn` is presently provided as a collection of research notebooks with limited documentation and without a consolidated, performance-oriented software architecture suitable for broader reuse. `TRAILS`, although methodologically advanced, currently relies on annual discretization and sequential year-specific calculations rather than a unified matrix-based integration of both dimensions.
+Existing dLCA tools such as `Temporalis` [@Cardellini:2018] handle temporal distribution but not temporal evolution. Conversely, pLCA tools like `premise` [@Sacchi:2022], `Futura` [@Joyce:2022], and `pathways` [@Sacchi:2024] model evolving systems but not temporal distributions within the supply chain. Two recent tools combine both temporal distribution and evolution: `ProsperDyn` [@LangQuantzendorff:2025] and `TRAILS` [@Sacchi:2026]. `ProsperDyn` is presently provided as a collection of research notebooks with limited documentation and without a consolidated, performance-oriented software architecture suitable for broader reuse. `TRAILS`, although methodologically advanced, currently relies on annual discretization and sequential year-specific calculations rather than a unified matrix-based integration of both dimensions.
 
 `bw_timex` uniquely embeds the time dimension directly into the technosphere and biosphere matrices, enabling flexible temporal resolution within a single matrix-based framework. This allows efficient computation and seamless integration with the broader `Brightway` ecosystem.
 
 # Workflow
 
-A time-explicit LCA with `bw_timex` follows four main steps, as illustrated in \autoref{fig:workflow}. First, a conventional product system model is temporalized by adding process-relative temporal distributions (rTDs) to the exchanges (cf. @Cardellini:2018). These rTDs describe how the amount of a technosphere or biosphere exchange is distributed over time, relative to the consuming or emitting process. In addition, temporal evolution of foreground processes can be defined through time-specific parameters. In Step 2, a timeline of technosphere exchanges is constructed by convolving rTDs along the supply chain, starting from the absolute reference time for the demand, which is defined by the user. In Step 3, the exchanges in the timeline are re-linked to time-specific background databases that reflect the technology landscape at specific points in time. Based on the temporally re-linked product system, a time-explicit LCI is calculated, preserving the timing of processes and emissions. The inventory is calculated following the conventional matrix-based LCA formulation [@Heijungs:2002], with the time dimension embedded in the matrices through additional row/column pairs. In Step 4, these emissions are characterized, either using standard characterization factors or by applying dynamic characterization functions that take the emissions’ timing into account.
[…]
