+++ environment.yml
@@ -6,7 +6,7 @@ channels:
   - diepers
 dependencies:
   - python==3.11
-  - bw_timex>=0.3.4
+  - bw_timex>=1.0.0
   # - brightway25_nosolver # uncomment for macOS ARM machines
   # - scikit-umfpack==0.3.3 # uncomment for macOS ARM machines
   # - numpy==1.24 # uncomment for macOS ARM machines
