### 2026-03-26 3ebeac2
add ltt abbreviation
