### 2026-03-26 #9621 @TimoDiepers
Thank you @arfon. I've released v1.0.0 and archived the code under https://doi.org/10.5281/zenodo.19230937.
