### 2026-04-13 #9621 @TimoDiepers
Hi @arfon, just checking back in if you need anything else to move this forward.

Version: v1.0.0
Archive: https://doi.org/10.5281/zenodo.19230937
