### 2026-04-16 93bb7db
add CITATION.cff
